pyproject.toml
[tool.poetry]
name = "bridgekeeper"
version = "0.7"
description = "Django permissions that work with QuerySets."
homepage = "https://bridgekeeper.readthedocs.io/"
repository = "https://github.com/excitedleigh/bridgekeeper"
authors = ["Leigh Brenecki <leigh@brenecki.id.au>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.5"

[tool.poetry.dev-dependencies]
djangorestframework = "^3.10"
pytest = "^5.0"
pytest-django = "^3.5"
pytest-pythonpath = "^0.7.3"
prospector = "^1.1"
# poetry seems to choke on this so for now, it's installed by nox instead
#factory-boy = "^2.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
